🏷️ Use the SoundEventId type from shared.ts, rather than having a separate frontend one.
